Add files via upload
Two files I forgot to upload. Stacked Histograms shows the MEC DIS RES and CCQE parts of the CC0pi spectrum in a stacked histogram. The Smear matrix shows the amounts of the total smear matrix that are formed of each of these types of scattering (which ended up showing that the RES and DIS channels contained the scattering with the most adverse effect on results).